requirements.txt
Django
djhtml
django-allauth[socialaccount]
django-allauth-ui
python-decouple
psycopg[binary]
dj-database-url
django-debug-toolbar
django-extensions
django-widget-tweaks
requests
faker
django-filter
django-htmx
pytest
pytest-django
factory-boy
django-template-partials
plotly
pandas
django-import-export
django-unfold
django-tailwind[reload]